SeekBar replaced with Slider

+++ app/src/main/java/ru/snakelord/philosofidget/data/datasource/settings/WidgetSettingsDataSource.kt
@@ -9,11 +9,11 @@ interface WidgetSettingsDataSource {
 
     fun setQuoteLanguage(language: String)
 
-    fun setQuoteTextSize(quoteTextSize: Int)
+    fun setQuoteTextSize(quoteTextSize: Float)
 
-    fun getQuoteTextSize(): Int
+    fun getQuoteTextSize(): Float
 
-    fun setQuoteAuthorTextSize(quoteAuthorTextSize: Int)
+    fun setQuoteAuthorTextSize(quoteAuthorTextSize: Float)
 
-    fun getQuoteAuthorTextSize(): Int
+    fun getQuoteAuthorTextSize(): Float
 }
